Console cleanup
- Remove webconsole wrapper
- Remove serviceContainer
