+++ src/devtools/client/webconsole/components/FilterBar/ConsoleSettings.js
@@ -132,7 +132,7 @@ class ConsoleSettings extends Component {
   render() {
     const { webConsoleUI } = this.props;
     const doc = webConsoleUI.document;
-    const toolbox = webConsoleUI.wrapper.toolbox;
+    const toolbox = webConsoleUI.toolbox;
 
     return MenuButton(
       {
