+++ src/devtools/client/webconsole/components/Output/message-types/EvaluationResult.js
@@ -16,7 +16,6 @@ EvaluationResult.propTypes = {
   dispatch: PropTypes.func.isRequired,
   message: PropTypes.object.isRequired,
   timestampsVisible: PropTypes.bool.isRequired,
-  serviceContainer: PropTypes.object,
   maybeScrollToBottom: PropTypes.func,
   open: PropTypes.bool,
 };
@@ -29,7 +28,6 @@ function EvaluationResult(props) {
   const {
     dispatch,
     message,
-    serviceContainer,
     timestampsVisible,
     maybeScrollToBottom,
     open,
@@ -66,7 +64,6 @@ function EvaluationResult(props) {
       dispatch,
       messageId,
       grip: parameters[0],
-      serviceContainer,
       useQuotes: true,
       escapeWhitespace: false,
       type,
@@ -85,7 +82,6 @@ function EvaluationResult(props) {
     topLevelClasses,
     messageBody,
     messageId,
-    serviceContainer,
     exceptionDocURL,
     stacktrace,
     collapsible: Array.isArray(stacktrace),
